feat(www): make the /create search an inline combobox
